Implemented experimental HLSL post-processing support on Direct3D 9 targets. [Ryan Holtz, Bat Country Entertainment] - You will need the latest DirectX redistributable from Microsoft's website. - You will need a new set of headers found here, including the caveats within: http://forums.bannister.org/ubbthreads.php?ubb=showflat&Number=69633 - 5-pass post-processing: Upscale, Post-Process, Store Last Frame, Defocus 1, Defocus 2 - Many tunable effects including: Scanlines, defocus, linear deconvergence, radial deconvergence, pincushion, RGB colorspace convolution, YIQ colorspace convolution, saturation, simulated dot crawl, simulated chroma subsampling, aperture masking, and more. - Requires a GPU that supports Shader Model 3.0 to be enabled and a powerful GPU, the entire pipeline consists of approximately 30 texel fetches and approximately 230 arthimetic ops. - Will supersample the framebuffer up to 9x in both X and Y, but this requires an enormously powerful GPU that has not been invented; users with Radeon 5000-class cards should limit themselves to 3x, Radeon 4000 to 1.5x. - The default configuration will NOT appear to do anything; it requires tuning to the user's liking. - Should nicely fall back in all cases except missing shaders, and it might fall back correctly in that case as well. Report any anomalies. - For obvious reasons, the Direct3D8 renderer cannot support this.
Artwork:
aperture.png: Based on Aaron Giles's aperture_4x6 mask but with more prominent notching to bring out the "brick" pattern that was used on many period colour monitors, including the Commodore 1702 - which was, admittedly, not an arcade monitor, but it looks believable and can be changed. It currently requires considerable fine-tuning to change masks but is doable.
white.png: Default 8x8 255,255,255,255 texture.

Initial shader import:
Focus: 8-sample blur that averages 7 samples around a center sample.
Phosphor: Not currently used, treated as a pass-through by drawd3d.c, but could be used to implement additional convolutions in a second pass.
Pincushion: Used (when commented in in drawd3d.c) to pincushion an entire full-screen texture but not otherwise apply any convolutions.
Post: The meat and potatoes. It does scanlines, it does aperture masking, it does dot crawl, it does chroma subsampling, it does YIQ colorspace convolution, it does RGB colorspace convolution, it does pincushioning, it walks, it talks, it does the dishes.
Primary: Simple passthrough for UI and artwork.